pyproject.toml
[tool.poetry]
name = "django-review-tdd"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "django_review_tdd"}]

[tool.poetry.dependencies]
python = ">=3.10"
django = ">=5.1.6,<6.0.0"
djangorestframework = ">=3.15.2,<4.0.0"
django-extensions = ">=3.2.3,<4.0.0"
pillow = ">=11.1.0,<12.0.0"
django-cleanup = ">=9.0.0,<10.0.0"
pymysql = ">=1.1.1,<2.0.0"
mysqlclient = ">=2.2.7,<3.0.0"
ipython = ">=8.32.0,<9.0.0"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"